Maven Setup Complete
I have the POM correctly created. Packaging 'app' creates a project that packages everything together and rund the javafx application starting in Environment.

I had to create a helper main because when I used Envoronment.java as the main class, the jar wouldnt run because there were 'missing javafx runtime components.'
I did some digging and found that this was a problem only if the mainClass was the javaFX application 'main,' and not if it was a class that didn't require any javafx. The helper class allowed essentially a workaround.