Added functionality to side panel
You can now click on a panel on the side and it will chadnge the main view so that it matched the selected currency.

I also realized that some of the currencies cannot be used as base(i think this is the problem) so I realized that I have to handle requests more safely(this was expected, my goal was to get basic functionality down and then to fix bugs to that it runs smoothly).